+++ app/src/main/kotlin/app/owlow/accsettings/ui/overview/OverviewRoute.kt
@@ -1,10 +1,11 @@
 package app.owlow.accsettings.ui.overview
 
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
+import androidx.lifecycle.Lifecycle
+import androidx.lifecycle.compose.LifecycleEventEffect
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.lifecycle.viewmodel.compose.viewModel
 
@@ -17,8 +18,12 @@ fun OverviewRoute(
 ) {
     val uiState by overviewViewModel.uiState.collectAsStateWithLifecycle()
 
-    LaunchedEffect(Unit) {
-        overviewViewModel.refresh().join()
+    LifecycleEventEffect(Lifecycle.Event.ON_START) {
+        overviewViewModel.startAutoRefresh()
+    }
+
+    LifecycleEventEffect(Lifecycle.Event.ON_STOP) {
+        overviewViewModel.stopAutoRefresh()
     }
 
     OverviewScreen(